organize resolvers and query utils
