+++ flow-typed/platform.js
@@ -1,7 +1,7 @@
 // @flow
 declare var Intl: Object;
 
-declare type Params = { [string]: string | number };
+declare type Params = { [string]: string };
 
 declare type FluxStandardAction = {
 	type: string,
@@ -62,29 +62,32 @@ declare type Match = {
 
 declare type MatchedRoute = { route: PlatformRoute, match: Match };
 declare type MatchPathOptions = {
-  path: string,
-  exact?: boolean,
-  strict?: boolean,
-}
+	path: string,
+	exact?: boolean,
+	strict?: boolean,
+};
 
 declare module 'react-router-dom/matchPath' {
-	declare function matchPath(pathname: string, options: MatchPathOptions): null | Match;
+	declare function matchPath(
+		pathname: string,
+		options: MatchPathOptions
+	): null | Match;
 	declare module.exports: typeof matchPath;
 }
 
 declare type LocationShape = {
-  pathname?: string,
-  search?: string,
-  hash?: string,
-  state?: any,
-}
+	pathname?: string,
+	search?: string,
+	hash?: string,
+	state?: any,
+};
 
 declare class RouterRedirect extends React$Component {
 	props: {
 		to: string | LocationShape,
 		push?: boolean,
-	}
+	},
 }
 declare module 'react-router-dom/Redirect' {
-	declare export default typeof RouterRedirect;
+	declare export default typeof RouterRedirect
 }
