+++ src/router/README.md
@@ -1,0 +1,3 @@
+# Router
+
+The Meetup Web Platform router module provides React routing utilities.
