+++ src/router/util/query.test.js
@@ -1,14 +1,7 @@
 import url from 'url';
-import { routes } from '../../tests/mockApp';
-import asyncRoutes from '../../tests/mockAsyncRoute';
-import {
-	activeRouteQueries,
-	decodeParams,
-	getChildRoutes,
-	resolveChildRoutes,
-	getRouteResolver,
-	getMatchedQueries,
-} from './routeUtils';
+import { routes } from '../../../tests/mockApp';
+import asyncRoutes from '../../../tests/mockAsyncRoute';
+import { activeRouteQueries, decodeParams, getMatchedQueries } from './query';
 
 describe('activeRouteQueries', () => {
 	it('gathers queries from nested routes', () => {
@@ -104,81 +97,6 @@ describe('decodeParams', () => {
 		expect(decoded).toEqual(object);
 	});
 });
-
-describe('getChildRoutes', () => {
-	it('returns [route.indexRoute] for exact match', () => {
-		const matchedRoute = {
-			route: { indexRoute: 'foo' },
-			match: { isExact: true },
-		};
-		const childRoutes = getChildRoutes(matchedRoute);
-		expect(childRoutes).toEqual([matchedRoute.route.indexRoute]);
-	});
-	it('returns route.routes for not-exact match', () => {
-		const matchedRoute = {
-			route: { indexRoute: 'foo', routes: ['bar', 'baz'] },
-			match: { isExact: false },
-		};
-		const childRoutes = getChildRoutes(matchedRoute);
-		expect(childRoutes).toEqual(matchedRoute.route.routes);
-	});
-	it('returns empty array for exact match without index route', () => {
-		const matchedRoute = {
-			route: { routes: ['bar', 'baz'] },
-			match: { isExact: true },
-		};
-		const childRoutes = getChildRoutes(matchedRoute);
-		expect(childRoutes).toEqual([]);
-	});
-	it('returns empty array for not-exact match without route.routes', () => {
-		const matchedRoute = {
-			route: { indexRoute: 'foo' },
-			match: { isExact: false },
-		};
-		const childRoutes = getChildRoutes(matchedRoute);
-		expect(childRoutes).toEqual([]);
-	});
-});
-
-describe('resolveChildRoutes', () => {
-	it('returns [route.indexRoute] for exact match', () => {
-		const matchedRoute = {
-			route: { indexRoute: 'foo' },
-			match: { isExact: true },
-		};
-		return resolveChildRoutes(matchedRoute).then(childRoutes => {
-			expect(childRoutes).toEqual([matchedRoute.route.indexRoute]);
-		});
-	});
-	it('returns route.routes for not-exact match', () => {
-		const matchedRoute = {
-			route: { indexRoute: 'foo', routes: ['bar', 'baz'] },
-			match: { isExact: false },
-		};
-		return resolveChildRoutes(matchedRoute).then(childRoutes => {
-			expect(childRoutes).toEqual(matchedRoute.route.routes);
-		});
-	});
-	it('returns empty array for exact match without index route', () => {
-		const matchedRoute = {
-			route: { routes: ['bar', 'baz'] },
-			match: { isExact: true },
-		};
-		return resolveChildRoutes(matchedRoute).then(childRoutes => {
-			expect(childRoutes).toEqual([]);
-		});
-	});
-	it('returns empty array for not-exact match without route.routes', () => {
-		const matchedRoute = {
-			route: { indexRoute: 'foo' },
-			match: { isExact: false },
-		};
-		return resolveChildRoutes(matchedRoute).then(childRoutes => {
-			expect(childRoutes).toEqual([]);
-		});
-	});
-});
-
 describe('getMatchedQueries', () => {
 	it('returns queries derived from the query function in matched route', () => {
 		const location = new URL('http://foo.com/bar/baz');
@@ -224,40 +142,6 @@ describe('getMatchedQueries', () => {
 		expect(matchedRoute.route.query).toHaveBeenCalledWith({
 			...match,
 			location,
-		});
-	});
-});
-
-describe('getRouteResolver', () => {
-	const bar = {
-		path: '/bar',
-	};
-	const foo = {
-		path: '/foo',
-		routes: [bar],
-	};
-	const qux = {
-		path: '/qux',
-	};
-	const baz = {
-		path: '/baz',
-		getNestedRoutes: () => Promise.resolve([qux]),
-	};
-	const root = {
-		path: '/',
-		routes: [foo, baz],
-	};
-	const resolveRoutes = getRouteResolver([root]);
-	it('returns all matched synchronous routes', () => {
-		const location = new URL('http://example.com/foo/bar');
-		return resolveRoutes(location).then(matchedRoutes => {
-			expect(matchedRoutes.map(({ route }) => route)).toEqual([root, foo, bar]);
-		});
-	});
-	it('returns all matched asynchronous routes', () => {
-		const location = new URL('http://example.com/baz/qux');
-		return resolveRoutes(location).then(matchedRoutes => {
-			expect(matchedRoutes.map(({ route }) => route)).toEqual([root, baz, qux]);
 		});
 	});
 });
