+++ src/router/util/resolve.js
@@ -18,15 +18,6 @@ const _routeMatchOptions = (
 	exact: route.exact,
 });
 
-export const decodeParams = (params: { [string]: any }): Params =>
-	Object.keys(params).reduce((decodedParams, key) => {
-		if (typeof params[key] !== 'undefined') {
-			// skip 'undefined' values that cannot be encoded (null is okay)
-			decodedParams[key] = params[key] && decodeURI(params[key]);
-		}
-		return decodedParams;
-	}, {});
-
 /*
  * Determine whether the indexRoute or nested route should be considered the
  * child route for a particular MatchedRoute
@@ -75,42 +66,6 @@ const _resolveIndexRoute: Resolver<MatchedRoute> = matchedRoute =>
 				.then(indexRoute => (matchedRoute.route.indexRoute = indexRoute))
 				.then(() => matchedRoute)
 		: Promise.resolve(matchedRoute);
-
-/*
- * find all routes from a given array of route config objects that match the
- * supplied `location`
- *
- * this function matches the signature of `react-router-config`'s `_resolveRouteMatches`
- * function, but interprets all `route.path` settings as nested
- *
- * @see {@link https://github.com/ReactTraining/react-router/tree/master/packages/react-router-config#matchroutesroutes-pathname}
- */
-/*
- * Collect the return values of each query functions associated with the matched
- * routes, called with the provided location URL
- *
- * The function returned from calling `_matchedRouteQueriesReducer` with a
- * `location` should be used as the callback to an `array.reduce` call.
- */
-const _matchedRouteQueriesReducer = (location: URL) => (
-	queries: Array<Query>,
-	{ route, match }: MatchedRoute
-): Array<Query> => {
-	if (!route.query) {
-		return queries;
-	}
-	const routeQueryFns = route.query instanceof Array
-		? route.query
-		: [route.query];
-
-	// call the query functions with non-url-encoded params
-	const params = decodeParams(match.params);
-	const routeQueries = routeQueryFns
-		.map(queryFn => queryFn({ ...match, location, params }))
-		.filter(query => query);
-
-	return [...queries, ...routeQueries];
-};
 
 /*
  * Find all routes in the routes array that match the provided URL, including
@@ -164,25 +119,4 @@ export const getRouteResolver = (
 ) => (location: URL): Promise<Array<MatchedRoute>> => {
 	const path = location.pathname.replace(baseUrl, '');
 	return _resolveRouteMatches(routes, path);
-};
-
-/*
- * A synchronous, curried interface to derive the query values returned by the
- * query functions of a provided set of routes given a particular location
- */
-export const getMatchedQueries = (location: URL) => (
-	matchedRoutes: Array<MatchedRoute>
-): Array<Query> =>
-	matchedRoutes.reduce(_matchedRouteQueriesReducer(location), []);
-
-/*
- * A curried interface into `_resolveRouteMatches` + `getMatchedQueries`
- */
-export const activeRouteQueries = (
-	routes: Array<PlatformRoute>,
-	baseUrl: string
-) => {
-	const resolveRoutes = getRouteResolver(routes, baseUrl);
-	return (location: URL) =>
-		resolveRoutes(location).then(getMatchedQueries(location));
 };
